Move initial home content to FileUpload, add logic for navigating to previous and next page
